+++ include/pow.hpp
@@ -7,13 +7,13 @@
 
 class POW : public FPR { 
   public:
-  std::vector<int> base(std::vector<int> n) {
+  std::vector<long int  > base(std::vector<long int  > n) {
     Z z;
     S s;
     return s * z(n); 
   }
 
-  std::vector<int> recurrence(std::vector<int> n) {
+  std::vector<long int  > recurrence(std::vector<long int  > n) {
     MULT mult;
     P p;
     return mult({p(1, n), p(2, n)});
